feat: Implement level 2 balance, content, and validation
This commit introduces a comprehensive set of changes to establish the Level 2 gameplay experience.

Features:
- Adds three new Level 2 monsters with distinct roles: Speed (Shadow Mantis), Power (Axe Goblin), and Normal (Giant Slime).
- Adds two new items crucial for L2 progression: the Iron Sword (equipment) and a Small Potion (usable item).
- Implements the logic for the Small Potion as a usable inventory item, adding a `handleUsePotion` handler.
- Creates `floor_02.json`, a new map populated with a strategic mix of L1 and L2 monsters and the new items.
- Updates player level progression data for L2 and L3, including a slower speed curve and balanced EXP requirements, based on user feedback.

Validation:
- Adds a powerful `gamedata-checker.ts` script to programmatically validate all new and existing balance constraints. This script was used to iterate on the design and ensure all numerical requirements from the issue were met precisely.
- Updates existing brittle tests in the checker to align with the new, correct game balance.

Documentation:
- Creates a detailed report (`jules/plan001-report.md`) of the entire process.
- Updates `jules.md` to document the new L2-L3 balance and the usage of the gamedata checker.
