+++ gamedata/monsters/level1_average.json
@@ -5,7 +5,7 @@
     "level": 1,
     "maxhp": 20,
     "attack": 16,
-    "defense": 6,
+    "defense": 3,
     "speed": 6,
     "gold": 10,
     "hp": 20,
