+++ jules.md
@@ -99,7 +99,7 @@
 
 ### 7.2 经验获取
 - **怪物奖励**: 玩家击败怪物后，会获得该怪物提供的经验值。
-- **计算公式**: `奖励经验 = (怪物MaxHP / 10) + 怪物攻击力 + 怪物防御力 + 怪物速度`
+- **计算公式**: `奖励经验 = floor(怪物MaxHP / 10) + 怪物攻击力 + 怪物防御力 + 怪物速度`
 
 ### 7.3 升级规则
 - **数据驱动**: 玩家的升级所需经验和升级后的属性，完全由 `gamedata/leveldata.json` 文件定义。该文件是一个等级表，详细列出了每个等级对应的属性值。
@@ -241,33 +241,40 @@ if (item) {
 
 ## 14. 游戏平衡性说明 (Game Balance Notes)
 
-### 14.1 等级1和等级2的数值设计 (Level 1 & 2 Stat Design)
+### 14.1 等级1的数值设计 (Level 1 Stat Design)
 
-为了建立游戏的初始挑战和玩家成长曲线，我们为玩家的前两个等级以及三种新的1级怪物（攻击型、防御型、平均型）设计了详细的数值。
+为了建立游戏的初始挑战和玩家成长曲线，我们为玩家的1级和三种1级怪物（攻击型、防御型、平均型）设计了详细的数值。核心设计目标是让玩家在1级时，面对7只特定怪物（2攻、2防、3平）的挑战，战斗结束后HP损失约80%，并刚好升级到2级。
+
+### 14.2 等级2到等级3的数值设计与校验工具 (Level 2 to 3 Balance and Validation Tool)
+
+在等级1的基础上，我们为玩家设计了更具挑战性的等级2体验，并成功开发和利用了自动化校验工具来保证其平衡性。
 
 **核心设计目标:**
-- 玩家在1级时，面对7只特定怪物（2攻、2防、3平）的挑战，战斗结束后HP损失约80%，并刚好升级。
-- 玩家升到2级后，属性有明显提升，但1级的怪物（特别是攻击型）依然能构成一定威胁。
+- 玩家在2级时，面对由1级和2级怪物混合组成的楼层，必须借助新获得的“铁剑”和“小回复药”才能成功清场。
+- 怪物伤害足够高，对玩家有明显威胁，总潜在伤害远超玩家2级时的最大HP。
+- 玩家升到3级后，属性有显著提升（特别是速度），能反超之前无法先手的速攻型怪物，但2级怪物对其依然有威胁。
+- 所有经验值、HP等核心数值都应为易于理解的整数。
 
 **最终实现的数值:**
 
 - **玩家等级数据 (`gamedata/leveldata.json`):**
-  - **等级 1:** `{ "maxhp": 150, "attack": 12, "defense": 10, "speed": 10 }`
-    - *注意: 玩家初始攻击力从10提升至12，这是为了满足“2回合击杀攻击型怪物”的设计要求。*
-  - **等级 2:** `{ "exp_needed": 220, "maxhp": 200, "attack": 17, "defense": 15, "speed": 15 }`
+  - **等级 2:** `{ "maxhp": 200, "attack": 17, "defense": 15, "speed": 12 }`
+  - **等级 3:** `{ "exp_needed": 710, "maxhp": 300, "attack": 25, "defense": 22, "speed": 18 }`
 
 - **新怪物数据 (`gamedata/monsters/`):**
-  - **攻击型 (`level1_offensive.json`):** `{ "maxhp": 20, "attack": 25, "defense": 2, "speed": 8 }`
-  - **平均型 (`level1_average.json`):** `{ "maxhp": 20, "attack": 16, "defense": 6, "speed": 6 }`
-  - **防御型 (`level1_defensive.json`):** `{ "maxhp": 30, "attack": 14, "defense": 7, "speed": 4 }`
+  - **速攻型 (暗影螳螂):** `{ "maxhp": 50, "attack": 53, "defense": 5, "speed": 15 }`
+  - **强攻型 (持斧哥布林):** `{ "maxhp": 60, "attack": 38, "defense": 10, "speed": 10 }`
+  - **普通型 (巨型史莱姆):** `{ "maxhp": 70, "attack": 33, "defense": 12, "speed": 5 }`
 
-### 14.2 数据校验工具的挑战 (Challenges with the Validation Tool)
+- **新道具数据:**
+  - **铁剑 (`equipments/iron_sword.json`):** `{ "attack": 13 }`
+  - **小回复药 (`items/small_potion.json`):** `{ "heal_amount": 80 }`
 
-在本次任务中，我们尝试为 `scripts/gamedata-checker.ts` 增加一个自动化校验模块，用于持续验证上述数值平衡的正确性。
-
-然而，在开发过程中遇到了一个与项目`ts-node`执行环境相关的顽固错误。该错误表现为，当校验脚本尝试访问从JSON加载的数据对象的任何属性时，脚本会立即崩溃，且不返回任何有意义的错误信息 (`[Object: null prototype] {}`)。
-
-经过大量的、系统的调试，我们能够将问题精确地定位到“在脚本中访问数据对象属性”这一行为上，但无法找到根本原因。这似乎是当前项目构建/执行环境中的一个深层问题。
-
-**结论:**
-为了避免项目进度被此技术难题无休止地拖延，我们决定**暂时搁置**自动化校验功能的开发。核心的游戏平衡性数据已经按计划实现并手动验证。未来的开发者在尝试扩展此校验脚本时，需要注意这个潜在的环境问题。
+**数据校验工具 (`scripts/gamedata-checker.ts`):**
+- **现状**: 该脚本已成功扩展，现在包含一个专门的`validateLevel2To3Balance`函数，用于全面校验上述所有设计约束。
+- **重要性**: 该工具是确保数值平衡正确性的核心。在开发过程中，它帮助我们快速定位并修复了多个因数值调整而产生的连锁问题，例如EXP总值计算错误、伤害排序错误等。之前遇到的执行环境问题已解决。
+- **如何使用**:
+  ```bash
+  npm run check-gamedata
+  ```
+  该命令会加载所有游戏数据并运行所有校验。如果任何平衡性约束被破坏，脚本将失败并报告详细错误。强烈建议在对`gamedata`目录下的任何文件进行修改后，都运行此脚本进行验证。
