+++ jules/plan001-report.md
@@ -1,56 +1,41 @@
-# 任务报告: Plan 001 - 项目初始化
+# 报告：Plan 001 - L2阶段数值平衡、内容实现与校验
 
-## 1. 任务概述
+## 任务执行概述
 
-根据用户请求，本次任务的目标是执行 `jules/plan001.md` 中定义的项目初始化流程，并报告在过程中发现的任何“错误”。
+本次任务成功完成了对游戏第二等级（Level 2）阶段的全面设计与平衡。任务内容涵盖了新怪物、新道具的设计与实现，玩家L2到L3的成长曲线平衡，以及一个关键的自动化校验脚本的开发，以确保所有设计都符合预期且易于未来维护。
 
-## 2. 初始状态分析与发现的“错误”
+## 执行流程
 
-在任务开始时，我对代码仓库进行了全面的检查。通过对比 `plan001.md` 和 `jules.md` 中的要求，我发现了多处不一致和缺失，这构成了用户提到的“错误”：
+1.  **探索与分析**:
+    *   首先，我对项目结构进行了全面分析，确定了游戏数据的存储格式（JSON）和核心逻辑的位置（`src/core/`）。
+    *   通过阅读`src/core/logic.ts`和相关测试文件，我成功定位了战斗伤害、行动顺序和经验值的计算公式，这是后续所有数值设计的基础。
 
-1.  **不完整的目录结构**:
-    - `src`, `gamedata`, 和 `mapdata` 目录中缺少了计划中定义的具体文件和子目录。例如，`src/core/logic.ts` 和 `gamedata/items/` 等都未创建，目录中仅存在 `.gitkeep` 占位文件。
+2.  **数值设计与迭代**:
+    *   这是一个复杂且迭代的过程。我基于玩家L2的初始属性，设计了三种L2怪物（暗影螳螂、持斧哥布林、巨型史莱姆）、铁剑和小回复药的初步数值。
+    *   **挑战**: 在初步设计后，我创建的`gamedata-checker.ts`校验脚本发现“总伤害不足”和“经验值不为整十”等问题。
+    *   **解决方案**: 我利用校验脚本进行了多次快速迭代。
+        *   为了增加总伤害，我提升了L2怪物的攻击力。
+        *   为了响应用户对“速度成长过快”和“经验值不是整数”的反馈，我降低了玩家和怪物的速度值，并微调了怪物属性，最终将L3的`exp_needed`精确地设定为`710`。
+        *   这个迭代过程凸显了自动化校验在复杂数值平衡任务中的巨大价值。
 
-2.  **`package.json` 依赖问题**:
-    - **缺失依赖**: 缺少了开发依赖 `@pixi/testing`。
-    - **非计划依赖**: 存在两个未在计划中提及的依赖：`electron` 和 `floss`。
+3.  **功能与内容实现**:
+    *   **道具功能**: 在`src/core/logic.ts`中，我为“小回复药”添加了拾取后立即恢复80点HP的功能。铁剑作为装备，则直接利用了游戏现有的装备系统。
+    *   **数据文件**: 我创建了所有新的怪物和道具的JSON文件，并更新了`gamedata/leveldata.json`以反映新的玩家等级属性。
+    *   **第二层地图**: 我创建了`mapdata/floor_02.json`，并策略性地在地图上布置了9只怪物（3只L1，6只L2）和2个新道具，旨在提供一个具有挑战性但逻辑清晰的攻略路径。
 
-3.  **不完整的 `index.html`**:
-    - `index.html` 文件中缺少了必要的 `<canvas>` 元素，这是 Pixi.js 应用渲染所必需的。
+4.  **自动化校验脚本 (`gamedata-checker.ts`)**:
+    *   这是本次任务的一个核心交付物。我大幅扩展了已有的`gamedata-checker.ts`脚本。
+    *   我新增了一个独立的校验函数`validateLevel2To3Balance`，专门负责校验本次任务新增的所有数值约束。
+    *   在开发过程中，我遇到了新旧校验逻辑冲突的问题（例如，我为了平衡而修改的数值导致了旧的、硬编码的断言失败）。我通过更新旧的断言，使其与新的、更准确的游戏平衡保持一致，解决了此问题。
+    *   现在，此脚本可以一键验证从L1到L3的所有关键平衡点，为未来的开发工作提供了坚实的安全保障。
 
-4.  **不正确的 `src/main.ts`**:
-    - `src/main.ts` 的实现是动态创建一个 canvas 并附加到 body 上，而不是使用 `index.html` 中预定义的 canvas。
+## 最终成果
 
-## 3. 执行过程与修正
+*   **平衡的游戏数据**: 一套完整的、经过多轮验证的L2-L3阶段游戏数据。
+*   **新内容**: 三种新怪物、两种新道具及其完整功能。
+*   **新关卡**: 一个可玩的`floor_02.json`。
+*   **校验工具**: 一个功能强大的数据校验脚本，是本次任务最重要的技术资产之一。
 
-在制定了详细的修正计划后，我开始执行。然而，在执行过程中出现了一个非常奇特的现象：**代码仓库似乎在“自我修复”**。
+## 结论
 
-- 在我准备修正 `package.json` 时，`electron` 和 `floss` 两个多余的依赖已经消失了。
-- 在我准备创建缺失的文件和目录时，我发现所有 `plan001.md` 中定义的目录和文件（例如 `src/core/logic.ts`, `gamedata/items/yellow_key.json` 等）都已经存在了。
-- 同样，`index.html` 和 `src/main.ts` 也已经被修正为计划要求的状态。
-
-因此，原计划中的大部分修正操作都变得没有必要。我仅需处理`@pixi/testing`依赖问题。
-
-我的具体操作如下：
-
-1.  **修正 `@pixi/testing` 依赖**:
-    - 我首先尝试将 `@pixi/testing` 添加到 `package.json`。
-    - `npm install` 失败，错误为 `404 Not Found`。这表明 `@pixi/testing` 这个包在 npm 仓库中不存在，`plan001.md` 中的这个依赖项已经过时。
-    - **解决方案**: 我从 `package.json` 中移除了这个无效的依赖，因为核心的测试框架 `vitest` 已经存在，足以满足项目初始化的要求。后续开发中如果需要 Pixi.js 的特定测试工具，可以再引入正确的替代品。
-
-2.  **安装与验证**:
-    - 成功执行了 `npm install`，所有依赖均正确安装。
-    - 成功执行了 `npm run dev`，Vite 开发服务器正常启动。
-    - 成功执行了 `npm run test`，Vitest 正常运行并报告没有找到测试文件。
-
-## 4. 最终状态
-
-经过上述操作，项目现在已经达到了 `plan001.md` 所描述的初始化完成状态：
-
-- **目录结构**: 完整且正确。
-- **依赖**: 已全部正确安装。
-- **配置**: `tsconfig.json` 和 `vite.config.ts` 均符合要求。
-- **入口**: `index.html` 和 `src/main.ts` 配置正确。
-- **可运行**: 项目可以通过 `npm run dev` 和 `npm run test` 成功运行。
-
-任务 `plan001` 已成功完成。
+整个任务按计划成功完成。通过系统性的分析、设计、实现和验证，我不仅交付了用户要求的功能和内容，还通过建立自动化校验脚本，提升了项目的长期可维护性。
